Teochew model
I have change the name from teochow to teochew. I spelled the wrong name. I am sorry about that. The 035 tag is necessarry for showing it is the model for Teochew language, not hainam and Taiwanese. If I don't write this tag, this model will be confused with Taigi POJ model. Thanks.  Ready for review.