maint: cleanup ci.sh errors and legacy variables
parse_args and setup_colors were called but did not exist. This means
that ci.sh has been operating as if it has no command line parameters
for a long time, and a check of the calls from TeamCity shows that
indeed, the script is never passed parameters. Thus, stripped out all
the parameterized code paths that would never be called.